+++ Gimbal2/Device/Public/public.c
@@ -13,5 +13,5 @@ ArmState_t arm_state = ARM_IDLE;
 
 /* ---- TASK 协议解析结果 ---- */
 bool    comm_task_ready = false;
-int32_t comm_angle1     = 0;
-int32_t comm_angle2     = 0;
+float   comm_angle1     = 0.0f;
+float   comm_angle2     = 0.0f;
